abschluss 10.2.2022
stand ohne Schere und Asterkennung

+++ LucanoV5/lib/lucano_lidar.h
@@ -5,7 +5,7 @@
 
 //      Prototypes
 void init_lidar1(void);
-void getTF_High_Data(int* distance, int* strength);
+void getTF_High_Data(uint16_t* distance, uint16_t* strength);
 void init_I2C_lidar();
 int getTFminiDataI2C(void);
